+++ src/vm/mem.c
@@ -135,10 +135,10 @@ mem_getStringLength(PmMemSpace_t memspace, uint8_t const *const pstr)
 
 
 PmReturn_t
-mem_cmpn(uint8_t *cname, uint8_t cnamelen, PmMemSpace_t memspace,
+mem_cmpn(uint8_t *cname, uint16_t cnamelen, PmMemSpace_t memspace,
          uint8_t const **paddr)
 {
-    uint8_t i;
+    uint16_t i;
     uint8_t b;
 
     /* Iterate over all characters */
